requirements.txt
datetime
uuid
flask
uwsgi
Flask-Login
pyOpenSSL
oauthlib
httplib2
google-api-python-client
requests
oauth2client
reload
Flask-Assets
cssmin
jsmin
pytest
redis
psycopg2-binary
sqlalchemy


